+++ financial-connections/src/main/java/com/stripe/android/financialconnections/features/institutionpicker/InstitutionPickerScreen.kt
@@ -1,9 +1,9 @@
-@file:Suppress("TooManyFunctions")
+@file:Suppress("TooManyFunctions", "LongMethod")
 
 package com.stripe.android.financialconnections.features.institutionpicker
 
 import androidx.activity.compose.BackHandler
-import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
@@ -34,6 +34,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.focus.onFocusChanged
+import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
@@ -49,17 +50,21 @@ import com.airbnb.mvrx.Uninitialized
 import com.airbnb.mvrx.compose.collectAsState
 import com.airbnb.mvrx.compose.mavericksViewModel
 import com.stripe.android.financialconnections.R
+import com.stripe.android.financialconnections.features.common.InstitutionPlaceholder
+import com.stripe.android.financialconnections.features.common.LoadingShimmerEffect
 import com.stripe.android.financialconnections.features.common.LoadingSpinner
 import com.stripe.android.financialconnections.features.institutionpicker.InstitutionPickerState.Payload
 import com.stripe.android.financialconnections.model.FinancialConnectionsInstitution
 import com.stripe.android.financialconnections.model.InstitutionResponse
 import com.stripe.android.financialconnections.presentation.parentViewModel
+import com.stripe.android.financialconnections.ui.LocalImageLoader
 import com.stripe.android.financialconnections.ui.TextResource
 import com.stripe.android.financialconnections.ui.components.AnnotatedText
 import com.stripe.android.financialconnections.ui.components.FinancialConnectionsOutlinedTextField
 import com.stripe.android.financialconnections.ui.components.FinancialConnectionsScaffold
 import com.stripe.android.financialconnections.ui.components.FinancialConnectionsTopAppBar
 import com.stripe.android.financialconnections.ui.theme.FinancialConnectionsTheme
+import com.stripe.android.uicore.image.StripeImage
 
 @Composable
 internal fun InstitutionPickerScreen() {
@@ -334,13 +339,16 @@ private fun InstitutionResultTile(
                 horizontal = 16.dp
             )
     ) {
-        Image(
-            painter = painterResource(id = R.drawable.stripe_ic_brandicon_institution),
+        val modifier = Modifier
+            .size(36.dp)
+            .clip(RoundedCornerShape(6.dp))
+        StripeImage(
+            url = institution.icon?.default ?: "",
+            imageLoader = LocalImageLoader.current,
             contentDescription = null,
-            modifier = Modifier
-                .size(36.dp)
-                .clip(RoundedCornerShape(6.dp))
-
+            modifier = modifier,
+            contentScale = ContentScale.Crop,
+            errorContent = { InstitutionPlaceholder(modifier) }
         )
         Spacer(modifier = Modifier.size(8.dp))
         Column {
@@ -389,19 +397,43 @@ private fun FeaturedInstitutionsGrid(
                         modifier = Modifier
                             .height(80.dp)
                             .fillMaxWidth()
+                            .clip(RoundedCornerShape(6.dp))
                             .border(
                                 width = 1.dp,
                                 color = FinancialConnectionsTheme.colors.borderDefault,
                                 shape = RoundedCornerShape(4.dp)
                             )
                             .clickable { onInstitutionSelected(institution) }
-                            .padding(8.dp)
                     ) {
-                        Text(
-                            text = institution.name,
-                            color = FinancialConnectionsTheme.colors.textPrimary,
-                            style = FinancialConnectionsTheme.typography.bodyEmphasized,
-                            textAlign = TextAlign.Center
+                        StripeImage(
+                            modifier = Modifier
+                                .fillMaxSize()
+                                .padding(8.dp),
+                            url = institution.logo?.default ?: "",
+                            imageLoader = LocalImageLoader.current,
+                            contentScale = ContentScale.Fit,
+                            loadingContent = {
+                                LoadingShimmerEffect { shimmer ->
+                                    Spacer(
+                                        modifier = Modifier
+                                            .align(Alignment.Center)
+                                            .height(20.dp)
+                                            .clip(RoundedCornerShape(10.dp))
+                                            .fillMaxWidth(fraction = 0.5f)
+                                            .background(shimmer)
+                                    )
+                                }
+                            },
+                            errorContent = {
+                                Text(
+                                    modifier = Modifier.align(Alignment.Center),
+                                    text = institution.name,
+                                    color = FinancialConnectionsTheme.colors.textPrimary,
+                                    style = FinancialConnectionsTheme.typography.bodyEmphasized,
+                                    textAlign = TextAlign.Center
+                                )
+                            },
+                            contentDescription = "Institution logo"
                         )
                     }
                 }