+++ paymentsheet-example/src/androidTestDebug/java/com/stripe/android/TestCustomers.kt
@@ -67,7 +67,7 @@ class TestCustomers {
         Billing.On,
         shipping = Shipping.Off,
         delayed = DelayedPMs.Off,
-        automatic = Automatic.On,
+        automatic = Automatic.Off,
         saveCheckboxValue = false,
         saveForFutureUseCheckboxVisible = false,
         useBrowser = Browser.Chrome,